requirements.txt
# Cat Caption Cage Match - Core Dependencies
# Install with: pip install -r requirements.txt

# Web UI framework
gradio>=4.0.0

# Google Gemini AI
google-generativeai>=0.8.0

# Database
duckdb>=1.0.0

# Image handling
pillow>=10.0.0
requests>=2.31.0

# Environment management
python-dotenv>=1.0.0
